fix(211-03): fix remaining Hardhat assertion mismatches (D-11)
- BurnieCoinflip: fix uint48->uint32 epoch packing in seed hash (2 tests)
- DegenerusAffiliate: update affiliateBonusPointsBest expected value 10->27
- DegenerusJackpots: fix traitBurnTicket slot 9->8 and ticketQueue slot 13->12
- DistressLootbox: fix pre-distress time buffer for day-based granularity

+++ test/unit/BurnieCoinflip.test.js
@@ -229,7 +229,7 @@ describe("BurnieCoinflip", function () {
       let found = false;
       for (let i = 0n; i < 10000n; i++) {
         const seed = hre.ethers.solidityPackedKeccak256(
-          ["uint256", "uint48"],
+          ["uint256", "uint32"],
           [i, epoch]
         );
         const seedBig = BigInt(seed);
@@ -250,7 +250,7 @@ describe("BurnieCoinflip", function () {
       let found = false;
       for (let i = 0n; i < 10000n; i++) {
         const seed = hre.ethers.solidityPackedKeccak256(
-          ["uint256", "uint48"],
+          ["uint256", "uint32"],
           [i, epoch]
         );
         const seedBig = BigInt(seed);
